Move tests to `test_committee_helpers.py` and `test_epoch_processing_helpers.py`
