+++ tests/eth2/beacon/test_helpers.py
@@ -1,23 +1,17 @@
 import copy
 import random
 
-import itertools
 import pytest
-
 from hypothesis import (
     given,
-    settings,
     strategies as st,
 )
 
 from eth_utils import (
-    big_endian_to_int,
     ValidationError,
 )
-
 from eth_utils.toolz import (
     assoc,
-    isdistinct,
 )
 
 from eth.constants import (
@@ -35,16 +29,10 @@ from eth2.beacon.constants import (
     GWEI_PER_ETH,
     FAR_FUTURE_EPOCH,
 )
-from eth2.beacon.exceptions import NoWinningRootError
 from eth2.beacon.enums import (
     SignatureDomain,
 )
-from eth2.beacon.state_machines.forks.serenity.blocks import (
-    SerenityBeaconBlock,
-)
-from eth2.beacon.types.attestations import (
-    Attestation,
-)
+
 from eth2.beacon.types.attestation_data import (
     AttestationData,
 )
@@ -52,21 +40,7 @@ from eth2.beacon.types.forks import Fork
 from eth2.beacon.types.slashable_attestations import SlashableAttestation
 from eth2.beacon.types.states import BeaconState
 from eth2.beacon.types.validator_records import ValidatorRecord
-from eth2.beacon.committee_helpers import (
-    get_beacon_proposer_index,
-    get_current_epoch_committee_count,
-    get_crosslink_committees_at_slot,
-    get_epoch_committee_count,
-    get_shuffling,
-    get_previous_epoch_committee_count,
-)
-from eth2.beacon.epoch_processing_helpers import (
-    get_attesting_validator_indices,
-    get_attestation_participants,
-    get_current_epoch_attestations,
-    get_previous_epoch_attestations,
-    get_winning_root,
-)
+
 from eth2.beacon.helpers import (
     _get_block_root,
     generate_aggregate_pubkeys,
@@ -88,16 +62,6 @@ import eth2._utils.bls as bls
 from tests.eth2.beacon.helpers import (
     get_pseudo_chain,
 )
-
-
-@pytest.fixture()
-def sample_block(sample_beacon_block_params):
-    return SerenityBeaconBlock(**sample_beacon_block_params)
-
-
-@pytest.fixture()
-def sample_state(sample_beacon_state_params):
-    return BeaconState(**sample_beacon_state_params)
 
 
 def generate_mock_latest_block_roots(
@@ -166,313 +130,6 @@ def test_get_block_root(current_slot,
             )
 
 
-#
-# Shuffling
-#
-@pytest.mark.parametrize(
-    (
-        'active_validator_count,'
-        'epoch_length,'
-        'target_committee_size,'
-        'shard_count,'
-        'expected_committee_count'
-    ),
-    [
-        (1000, 20, 10, 50, 40),  # SHARD_COUNT // EPOCH_LENGTH
-        (500, 20, 10, 100, 40),  # active_validator_count // EPOCH_LENGTH // TARGET_COMMITTEE_SIZE
-        (20, 10, 3, 10, 10),  # 1
-        (20, 10, 3, 5, 10),  # 1
-        (40, 5, 10, 2, 5),  # 1
-    ],
-)
-def test_get_epoch_committee_count(
-        active_validator_count,
-        epoch_length,
-        target_committee_size,
-        shard_count,
-        expected_committee_count):
-    assert expected_committee_count == get_epoch_committee_count(
-        active_validator_count=active_validator_count,
-        shard_count=shard_count,
-        epoch_length=epoch_length,
-        target_committee_size=target_committee_size,
-    )
-
-
-@pytest.mark.parametrize(
-    (
-        'num_validators,'
-        'epoch_length,'
-        'target_committee_size,'
-        'shard_count,'
-        'epoch'
-    ),
-    [
-        (1000, 20, 10, 100, 0),
-        (1000, 20, 10, 100, 0),
-        (1000, 20, 10, 100, 1),
-        (20, 10, 3, 10, 0),  # active_validators_size < epoch_length * target_committee_size
-        (20, 10, 3, 10, 0),
-        (20, 10, 3, 10, 1),
-    ],
-)
-def test_get_shuffling_is_complete(activated_genesis_validators,
-                                   epoch_length,
-                                   target_committee_size,
-                                   shard_count,
-                                   epoch):
-    shuffling = get_shuffling(
-        seed=b'\x35' * 32,
-        validators=activated_genesis_validators,
-        epoch=epoch,
-        epoch_length=epoch_length,
-        target_committee_size=target_committee_size,
-        shard_count=shard_count,
-    )
-
-    assert len(shuffling) == epoch_length
-    assert len(shuffling) > 0
-    for committee in shuffling:
-        assert len(committee) <= target_committee_size
-        assert len(committee) > 0
-    validator_indices = tuple(
-        itertools.chain(
-            [
-                validator_index
-                for committee in shuffling
-                for validator_index in committee
-            ]
-        )
-    )
-    assert isdistinct(validator_indices)
-    activated_genesis_validator_indices = tuple(
-        index
-        for index in range(len(activated_genesis_validators))
-    )
-    assert sorted(validator_indices) == sorted(activated_genesis_validator_indices)
-
-
-@pytest.mark.parametrize(
-    (
-        'n, target_committee_size, shard_count, len_active_validators,'
-        'previous_calculation_epoch, current_calculation_epoch,'
-        'get_prev_or_cur_epoch_committee_count,'
-        'delayed_activation_epoch'
-    ),
-    [
-        (
-            100, 10, 20, 20,
-            5, 10,
-            get_previous_epoch_committee_count,
-            5 + 1,
-        ),
-        (
-            100, 10, 20, 100,
-            5, 10,
-            get_previous_epoch_committee_count,
-            5 + 1,
-        ),
-        (
-            100, 10, 20, 20,
-            5, 10,
-            get_current_epoch_committee_count,
-            10 + 1,
-        ),
-        (
-            100, 10, 20, 100,
-            5, 10,
-            get_current_epoch_committee_count,
-            10 + 1,
-        ),
-    ],
-)
-def test_get_prev_or_cur_epoch_committee_count(
-        monkeypatch,
-        n_validators_state,
-        epoch_length,
-        n,
-        target_committee_size,
-        shard_count,
-        len_active_validators,
-        previous_calculation_epoch,
-        current_calculation_epoch,
-        get_prev_or_cur_epoch_committee_count,
-        delayed_activation_epoch):
-    from eth2.beacon import committee_helpers
-
-    def mock_get_epoch_committee_count(
-            active_validator_count,
-            shard_count,
-            epoch_length,
-            target_committee_size):
-        return active_validator_count // shard_count
-
-    monkeypatch.setattr(
-        committee_helpers,
-        'get_epoch_committee_count',
-        mock_get_epoch_committee_count
-    )
-
-    state = n_validators_state.copy(
-        slot=0,
-        previous_calculation_epoch=previous_calculation_epoch,
-        current_calculation_epoch=current_calculation_epoch,
-    )
-    for index in range(len(state.validator_registry)):
-        if index < len_active_validators:
-            validator = state.validator_registry[index].copy(
-                activation_epoch=0,
-            )
-            state = state.update_validator_registry(
-                index,
-                validator,
-            )
-        else:
-            validator = state.validator_registry[index].copy(
-                activation_epoch=delayed_activation_epoch,
-            )
-            state = state.update_validator_registry(
-                index,
-                validator,
-            )
-
-    result_committee_count = get_prev_or_cur_epoch_committee_count(
-        state=state,
-        shard_count=shard_count,
-        epoch_length=epoch_length,
-        target_committee_size=target_committee_size,
-    )
-    expected_committee_count = len_active_validators // shard_count
-
-    assert result_committee_count == expected_committee_count
-
-
-@pytest.mark.parametrize(
-    (
-        'current_slot,'
-        'slot,'
-        'epoch_length,'
-        'target_committee_size,'
-        'shard_count,'
-    ),
-    [
-        # genesis_epoch == previous_epoch == slot_to_epoch(slot) == current_epoch
-        (0, 5, 10, 10, 10),
-        # genesis_epoch == previous_epoch == slot_to_epoch(slot) < current_epoch
-        (10, 5, 10, 10, 10),
-        # genesis_epoch < previous_epoch == slot_to_epoch(slot) < current_epoch
-        (20, 11, 10, 10, 10),
-        # genesis_epoch == previous_epoch < slot_to_epoch(slot) == current_epoch
-        (10, 11, 10, 10, 10),
-    ],
-)
-def test_get_crosslink_committees_at_slot(
-        n_validators_state,
-        current_slot,
-        slot,
-        epoch_length,
-        target_committee_size,
-        shard_count,
-        genesis_epoch):
-
-    state = n_validators_state.copy(
-        slot=current_slot,
-    )
-
-    crosslink_committees_at_slot = get_crosslink_committees_at_slot(
-        state=state,
-        slot=slot,
-        genesis_epoch=genesis_epoch,
-        epoch_length=epoch_length,
-        target_committee_size=target_committee_size,
-        shard_count=shard_count,
-    )
-    assert len(crosslink_committees_at_slot) > 0
-    for crosslink_committee in crosslink_committees_at_slot:
-        committee, shard = crosslink_committee
-        assert len(committee) > 0
-        assert shard < shard_count
-
-
-#
-# Get proposer postition
-#
-@pytest.mark.parametrize(
-    (
-        'num_validators,'
-        'epoch_length,'
-        'committee,'
-        'slot,'
-        'success,'
-    ),
-    [
-        (
-            100,
-            64,
-            (10, 11, 12),
-            0,
-            True,
-        ),
-        (
-            100,
-            64,
-            (),
-            0,
-            False,
-        ),
-    ]
-)
-def test_get_beacon_proposer_index(
-        monkeypatch,
-        num_validators,
-        epoch_length,
-        committee,
-        slot,
-        success,
-        sample_state,
-        genesis_epoch,
-        target_committee_size,
-        shard_count):
-
-    from eth2.beacon import committee_helpers
-
-    def mock_get_crosslink_committees_at_slot(state,
-                                              slot,
-                                              genesis_epoch,
-                                              epoch_length,
-                                              target_committee_size,
-                                              shard_count):
-        return (
-            (committee, 1,),
-        )
-
-    monkeypatch.setattr(
-        committee_helpers,
-        'get_crosslink_committees_at_slot',
-        mock_get_crosslink_committees_at_slot
-    )
-    if success:
-        proposer_index = get_beacon_proposer_index(
-            sample_state,
-            slot,
-            genesis_epoch,
-            epoch_length,
-            target_committee_size,
-            shard_count,
-        )
-        assert proposer_index == committee[slot % len(committee)]
-    else:
-        with pytest.raises(ValidationError):
-            get_beacon_proposer_index(
-                sample_state,
-                slot,
-                genesis_epoch,
-                epoch_length,
-                target_committee_size,
-                shard_count,
-            )
-
-
 def test_get_active_validator_indices(sample_validator_record_params):
     current_epoch = 1
     # 3 validators are ACTIVE
@@ -499,420 +156,6 @@ def test_get_active_validator_indices(sample_validator_record_params):
     )
     active_validator_indices = get_active_validator_indices(validators, current_epoch)
     assert len(active_validator_indices) == 1
-
-
-@pytest.mark.parametrize(
-    (
-        'num_validators,'
-        'epoch_length,'
-        'committee,'
-        'aggregation_bitfield,'
-        'expected'
-    ),
-    [
-        (
-            100,
-            64,
-            (10, 11, 12),
-            b'\00',
-            (),
-        ),
-        (
-            100,
-            64,
-            (10, 11, 12),
-            b'\x80',
-            (10,),
-        ),
-        (
-            100,
-            64,
-            (10, 11, 12),
-            b'\xc0',
-            (10, 11),
-        ),
-        (
-            100,
-            64,
-            (10, 11, 12),
-            b'\x00\x00',
-            ValueError(),
-        ),
-    ]
-)
-def test_get_attestation_participants(
-        monkeypatch,
-        num_validators,
-        epoch_length,
-        committee,
-        aggregation_bitfield,
-        expected,
-        sample_state,
-        genesis_epoch,
-        target_committee_size,
-        shard_count,
-        sample_attestation_data_params):
-    shard = 1
-
-    from eth2.beacon import committee_helpers
-
-    def mock_get_crosslink_committees_at_slot(state,
-                                              slot,
-                                              genesis_epoch,
-                                              epoch_length,
-                                              target_committee_size,
-                                              shard_count):
-        return (
-            (committee, shard,),
-        )
-
-    monkeypatch.setattr(
-        committee_helpers,
-        'get_crosslink_committees_at_slot',
-        mock_get_crosslink_committees_at_slot
-    )
-
-    attestation_data = AttestationData(**sample_attestation_data_params).copy(
-        slot=0,
-        shard=shard,
-    )
-    if isinstance(expected, Exception):
-        with pytest.raises(ValidationError):
-            get_attestation_participants(
-                state=sample_state,
-                attestation_data=attestation_data,
-                bitfield=aggregation_bitfield,
-                genesis_epoch=genesis_epoch,
-                epoch_length=epoch_length,
-                target_committee_size=target_committee_size,
-                shard_count=shard_count,
-            )
-    else:
-        result = get_attestation_participants(
-            state=sample_state,
-            attestation_data=attestation_data,
-            bitfield=aggregation_bitfield,
-            genesis_epoch=genesis_epoch,
-            epoch_length=epoch_length,
-            target_committee_size=target_committee_size,
-            shard_count=shard_count,
-        )
-
-        assert result == expected
-
-
-@settings(max_examples=1)
-@given(random=st.randoms())
-@pytest.mark.parametrize(
-    (
-        'target_committee_size,'
-        'shard_count'
-    ),
-    [
-        (
-            16,
-            32,
-        ),
-    ]
-)
-def test_get_attesting_validator_indices(
-        random,
-        monkeypatch,
-        target_committee_size,
-        shard_count,
-        genesis_epoch,
-        epoch_length,
-        sample_state,
-        sample_attestation_data_params,
-        sample_attestation_params):
-    shard = 1
-    committee = tuple([i for i in range(target_committee_size)])
-
-    from eth2.beacon import committee_helpers
-
-    def mock_get_crosslink_committees_at_slot(state,
-                                              slot,
-                                              genesis_epoch,
-                                              epoch_length,
-                                              target_committee_size,
-                                              shard_count):
-        return (
-            (committee, shard,),
-        )
-
-    monkeypatch.setattr(
-        committee_helpers,
-        'get_crosslink_committees_at_slot',
-        mock_get_crosslink_committees_at_slot
-    )
-
-    # Validators attesting to two shard block roots
-    shard_block_root_1 = hash_eth2(b'shard_block_root_1')
-    shard_block_root_2 = hash_eth2(b'shard_block_root_2')
-
-    # Random sampling half the committee.
-    # `attestation_participants_1` and `attestation_participants_2` are expected to have
-    # overlapping participants.
-    attestation_participants_1 = random.sample(committee, target_committee_size // 2)
-    attestation_participants_2 = random.sample(committee, target_committee_size // 2)
-    not_attestation_participants_1 = [i for i in committee if i not in attestation_participants_1]
-
-    # Generate bitfield of each participants set
-    aggregation_bitfield_1 = get_empty_bitfield(target_committee_size)
-    aggregation_bitfield_2 = get_empty_bitfield(target_committee_size)
-    not_aggregation_bitfield_1 = get_empty_bitfield(target_committee_size)
-    for committee_index_1, committee_index_2, committee_index_3 in zip(
-            attestation_participants_1, attestation_participants_2, not_attestation_participants_1):
-        aggregation_bitfield_1 = set_voted(aggregation_bitfield_1, committee_index_1)
-        aggregation_bitfield_2 = set_voted(aggregation_bitfield_2, committee_index_2)
-        not_aggregation_bitfield_1 = set_voted(not_aggregation_bitfield_1, committee_index_3)
-
-    # `attestions` contains attestation to different block root by different set of participants
-    attestations = [
-        # Attestation to `shard_block_root_1` by `attestation_participants_1`
-        Attestation(**sample_attestation_params).copy(
-            data=AttestationData(**sample_attestation_data_params).copy(
-                shard=shard,
-                shard_block_root=shard_block_root_1,
-            ),
-            aggregation_bitfield=aggregation_bitfield_1
-        ),
-        # Attestation to `shard_block_root_1` by `attestation_participants_2`
-        Attestation(**sample_attestation_params).copy(
-            data=AttestationData(**sample_attestation_data_params).copy(
-                shard=shard,
-                shard_block_root=shard_block_root_1,
-            ),
-            aggregation_bitfield=aggregation_bitfield_2
-        ),
-        # Attestation to `shard_block_root_2` by `not_attestation_participants_1`
-        Attestation(**sample_attestation_params).copy(
-            data=AttestationData(**sample_attestation_data_params).copy(
-                shard=shard,
-                shard_block_root=shard_block_root_2,
-            ),
-            aggregation_bitfield=not_aggregation_bitfield_1
-        ),
-    ]
-
-    shard_block_root_1_attesting_validator = get_attesting_validator_indices(
-        state=sample_state,
-        attestations=attestations,
-        shard=shard,
-        shard_block_root=shard_block_root_1,
-        genesis_epoch=genesis_epoch,
-        epoch_length=epoch_length,
-        target_committee_size=target_committee_size,
-        shard_count=shard_count,
-    )
-    # Check that result is the union of two participants set
-    # `attestation_participants_1` and `attestation_participants_2`
-    assert set(shard_block_root_1_attesting_validator) == set(
-        attestation_participants_1 + attestation_participants_2)
-    assert len(shard_block_root_1_attesting_validator) == len(
-        set(attestation_participants_1 + attestation_participants_2))
-
-    shard_block_root_2_attesting_validator = get_attesting_validator_indices(
-        state=sample_state,
-        attestations=attestations,
-        shard=shard,
-        shard_block_root=shard_block_root_2,
-        genesis_epoch=genesis_epoch,
-        epoch_length=epoch_length,
-        target_committee_size=target_committee_size,
-        shard_count=shard_count,
-    )
-    # Check that result is the `not_attestation_participants_1` set
-    assert set(shard_block_root_2_attesting_validator) == set(not_attestation_participants_1)
-    assert len(shard_block_root_2_attesting_validator) == len(not_attestation_participants_1)
-
-
-@settings(max_examples=1)
-@given(random=st.randoms())
-def test_get_current_and_previous_epoch_attestations(random,
-                                                     sample_state,
-                                                     genesis_epoch,
-                                                     epoch_length,
-                                                     sample_attestation_data_params,
-                                                     sample_attestation_params):
-    num_previous_epoch_attestation, num_current_epoch_attestation = random.sample(
-        range(epoch_length),
-        2,
-    )
-    previous_epoch_attestion_slots = random.sample(
-        range(epoch_length),
-        num_previous_epoch_attestation,
-    )
-    current_epoch_attestion_slots = random.sample(
-        range(epoch_length, epoch_length * 2),
-        num_current_epoch_attestation,
-    )
-
-    previous_epoch_attestations = []
-    for slot in previous_epoch_attestion_slots:
-        previous_epoch_attestations.append(
-            Attestation(**sample_attestation_params).copy(
-                data=AttestationData(**sample_attestation_data_params).copy(
-                    slot=slot,
-                ),
-            )
-        )
-    current_epoch_attestations = []
-    for slot in current_epoch_attestion_slots:
-        current_epoch_attestations.append(
-            Attestation(**sample_attestation_params).copy(
-                data=AttestationData(**sample_attestation_data_params).copy(
-                    slot=slot,
-                ),
-            )
-        )
-
-    state = sample_state.copy(
-        slot=(epoch_length * 2 - 1),
-        latest_attestations=(previous_epoch_attestations + current_epoch_attestations),
-    )
-    assert set(previous_epoch_attestations) == set(
-        get_previous_epoch_attestations(state, epoch_length, genesis_epoch))
-    assert set(current_epoch_attestations) == set(
-        get_current_epoch_attestations(state, epoch_length))
-
-
-@settings(max_examples=10)
-@given(random=st.randoms())
-@pytest.mark.parametrize(
-    (
-        'target_committee_size,'
-        'block_root_1_participants,'
-        'block_root_2_participants,'
-    ),
-    [
-        (
-            16,
-            (1, 3),
-            (2, 4, 6, 8)
-        ),
-        (
-            16,
-            # vote tie; lower root value is favored
-            (1, 3, 5, 7),
-            (2, 4, 6, 8)
-        ),
-        (
-            16,
-            # no votes; no winning root
-            (),
-            ()
-        ),
-    ]
-)
-def test_get_winning_root(
-        random,
-        monkeypatch,
-        target_committee_size,
-        block_root_1_participants,
-        block_root_2_participants,
-        config,
-        n_validators_state,
-        sample_attestation_data_params,
-        sample_attestation_params):
-    shard = 1
-    committee = tuple([i for i in range(target_committee_size)])
-
-    from eth2.beacon import committee_helpers
-
-    def mock_get_crosslink_committees_at_slot(state,
-                                              slot,
-                                              genesis_epoch,
-                                              epoch_length,
-                                              target_committee_size,
-                                              shard_count):
-        return (
-            (committee, shard,),
-        )
-
-    monkeypatch.setattr(
-        committee_helpers,
-        'get_crosslink_committees_at_slot',
-        mock_get_crosslink_committees_at_slot
-    )
-
-    competing_block_roots = [
-        hash_eth2(bytearray(random.getrandbits(8) for _ in range(10))),
-        hash_eth2(bytearray(random.getrandbits(8) for _ in range(10)))
-    ]
-
-    # Generate bitfield of each participants set
-    root_1_participants_bitfield = get_empty_bitfield(target_committee_size)
-    root_2_participants_bitfield = get_empty_bitfield(target_committee_size)
-    for i in block_root_1_participants:
-        root_1_participants_bitfield = set_voted(root_1_participants_bitfield, i)
-    for i in block_root_2_participants:
-        root_2_participants_bitfield = set_voted(root_2_participants_bitfield, i)
-
-    # `attestions` contains attestation to different block root by different set of participants
-    attestations = [
-        # Attestation to `shard_block_root_1` by `attestation_participants_1`
-        Attestation(**sample_attestation_params).copy(
-            data=AttestationData(**sample_attestation_data_params).copy(
-                shard=shard,
-                shard_block_root=competing_block_roots[0],
-            ),
-            aggregation_bitfield=root_1_participants_bitfield
-        ),
-        # Attestation to `shard_block_root_2` by `attestation_participants_2`
-        Attestation(**sample_attestation_params).copy(
-            data=AttestationData(**sample_attestation_data_params).copy(
-                shard=shard,
-                shard_block_root=competing_block_roots[1],
-            ),
-            aggregation_bitfield=root_2_participants_bitfield
-        ),
-    ]
-
-    try:
-        winning_root, attesting_balance = get_winning_root(
-            state=n_validators_state,
-            shard=shard,
-            attestations=attestations,
-            genesis_epoch=config.GENESIS_EPOCH,
-            epoch_length=config.EPOCH_LENGTH,
-            max_deposit_amount=config.MAX_DEPOSIT_AMOUNT,
-            target_committee_size=config.TARGET_COMMITTEE_SIZE,
-            shard_count=config.SHARD_COUNT,
-        )
-        attesting_validators_indices = get_attesting_validator_indices(
-            state=n_validators_state,
-            attestations=attestations,
-            shard=shard,
-            shard_block_root=winning_root,
-            genesis_epoch=config.GENESIS_EPOCH,
-            epoch_length=config.EPOCH_LENGTH,
-            target_committee_size=config.TARGET_COMMITTEE_SIZE,
-            shard_count=config.SHARD_COUNT,
-        )
-        total_attesting_balance = sum(
-            get_effective_balance(
-                n_validators_state.validator_balances,
-                i,
-                config.MAX_DEPOSIT_AMOUNT
-            )
-            for i in attesting_validators_indices
-        )
-        assert attesting_balance == total_attesting_balance
-    except NoWinningRootError:
-        assert len(block_root_1_participants) == 0 and len(block_root_2_participants) == 0
-    else:
-        if len(block_root_1_participants) == len(block_root_2_participants):
-            root_1_as_int = big_endian_to_int(competing_block_roots[0])
-            root_2_as_int = big_endian_to_int(competing_block_roots[1])
-            if root_1_as_int < root_2_as_int:
-                assert winning_root == competing_block_roots[0]
-            else:
-                assert winning_root == competing_block_roots[1]
-        elif len(block_root_1_participants) < len(block_root_2_participants):
-            assert winning_root == competing_block_roots[1]
-        else:
-            assert winning_root == competing_block_roots[0]
 
 
 @pytest.mark.parametrize(
